include module tracing only if evm-tracing feature is enabled
